# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "IFG Lab Manager API",
        "description": "API RESTful para gerenciamento de laboratórios de informática do IFG",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor de API"
        }
    ],
    "paths": {
...
        "/api/v1/labs": {
            "get": {
                "tags": [
                    "Laboratórios"
                ],
                "summary": "Listar laboratórios",
                "operationId": "fe49face39a699dc5449b344354c3491",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de laboratórios"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Laboratórios"
                ],
                "summary": "Criar laboratório",
                "operationId": "60dc0ed42701f528ad8a0bf08b1a8111",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laboratório de Informática"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição do laboratório",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Laboratório criado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/labs/{id}": {
            "get": {
                "tags": [
                    "Laboratórios"
                ],
                "summary": "Obter detalhes do laboratório",
                "operationId": "09c552382b93569d03d91174aa14b0eb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do laboratório"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Laboratório não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/labs/{id}/computers": {
            "get": {
                "tags": [
                    "Laboratórios"
                ],
                "summary": "Listar computadores do laboratório",
                "description": "Get paginated computers for a lab",
                "operationId": "b391ddbf0c65b3c82dbdba423a137a4a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "online",
                                "offline"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de computadores"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/labs/{id}/softwares": {
            "get": {
                "tags": [
                    "Laboratórios"
                ],
                "summary": "Listar softwares do laboratório",
                "description": "Get paginated unique softwares for a lab",
                "operationId": "211649a52c54188b3b2bf7746a7561df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de softwares"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
}